requirements.txt
tensorflow==2.16.2
numpy==1.26.4
matplotlib==3.6.2
Pillow==9.5.0
fastapi==0.95.0